next step in change from phaseMask to phaseIds
change column phaseId to phaseIds to take more ids
translate all phaseMask to phaseIds.. to free column phaseMask
add random and duration files,
re-run from cmake is neccassary..

+++ src/server/game/DataStores/DBCStores.cpp
@@ -1317,7 +1317,7 @@ uint32 GetDefaultMapLight(uint32 mapId)
     return 0;
 }
 
-std::set<uint32> const& GetPhasesForGroup(uint32 group)
+std::set<uint16> const& GetPhasesForGroup(uint16 group)
 {
     return sPhasesByGroup[group];
 }
